requirements.txt
asgiref==3.8.1
asttokens==3.0.0
beautifulsoup4==4.13.4
colorama==0.4.6
decorator==5.2.1
Django==5.2
django-extensions==4.1
django-meta==2.5.0
django-robots==6.1
django-tinymce==4.1.0
executing==2.2.0
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
parso==0.8.4
pillow==11.2.1
prompt_toolkit==3.0.51
pure_eval==0.2.3
Pygments==2.19.1
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
wcwidth==0.2.13
Werkzeug==3.1.3
